Add Docker based validator
This commit adds a Docker file that can be run to validate any deploy.json.

The docker file looks at the declared version of the deploy.json, downloads the
corresponding schema file and runs a validation. If it is not valid, the process
will exit with a non-zero exit code.